+++ src/world/Objects/Object.h
@@ -631,7 +631,7 @@ class SERVER_DECL Object : public EventableObject, public IUpdatable
         /// \return none
         ///
         //////////////////////////////////////////////////////////////////////////////////////////
-        virtual void OutPacket(uint16 opcode, uint16 len, const void* data) {};
+        virtual void OutPacket(uint32 opcode, uint16 len, const void* data) {};
 
 
         //////////////////////////////////////////////////////////////////////////////////////////
@@ -648,7 +648,7 @@ class SERVER_DECL Object : public EventableObject, public IUpdatable
 
         virtual void SendMessageToSet(WorldPacket* data, bool self, bool myteam_only = false);
         void SendMessageToSet(StackBufferBase* data, bool self) { OutPacketToSet(data->GetOpcode(), static_cast<uint16>(data->GetSize()), data->GetBufferPointer(), self); }
-        virtual void OutPacketToSet(uint16 Opcode, uint16 Len, const void* Data, bool self);
+        virtual void OutPacketToSet(uint32 Opcode, uint16 Len, const void* Data, bool self);
 
         //////////////////////////////////////////////////////////////////////////////////////////
         ///void SendAIReaction(uint32 reaction = 2)
