+++ src/world/Spell/Spell.cpp
@@ -5863,7 +5863,7 @@ void Spell::SendCastSuccess(const uint64 & guid)
     *(uint32*)&buffer[c] = GetProto()->Id;
     c += 4;
 
-    plr->GetSession()->OutPacket(uint16(SMSG_CLEAR_EXTRA_AURA_INFO_OBSOLETE), static_cast<uint16>(c), buffer);
+    plr->GetSession()->OutPacket(uint32(SMSG_CLEAR_EXTRA_AURA_INFO_OBSOLETE), static_cast<uint16>(c), buffer);
 }
 
 uint8 Spell::GetErrorAtShapeshiftedCast(SpellEntry* spellInfo, uint32 form)
